Revert "宝宝心里苦啊……重新写了函数"
This reverts commit d00fc381e4f36e104f474198019a41a5d1835f69.

+++ Strategy/Strategy.h
@@ -24,7 +24,7 @@ int TreePolicy(int i);
 
 int Expand(int i);
 
-int BestChild(int v, double c);
+int BestChild(int v, int c);
 
 double DefaultPolicy(int v);
